+++ apps/dataviz-scrolly/src/App.tsx
@@ -1,0 +1,7 @@
+import { DataVizScrollyDemo } from "./components/DataVizScrollyDemo";
+
+function App() {
+  return <DataVizScrollyDemo />;
+}
+
+export default App;